requirements.txt
langchain-astradb
langchain 
datasets
pypdf
python-dotenv
flask
langchain_huggingface
faiss-cpu
-e .